refactor(shared): split phase_quality.py (1108 LOC) into a 9-module package + add bloat-findings column to Compliance Dashboard (Campaign B B3)
Iterate Campaign B sub-iterate B3 — campaign
`2026-05-25-bloat-cleanup-B-shipwright`. Splits the 1108-LOC
`shared/scripts/lib/phase_quality.py` monolith into a focused package
where every submodule is well under the 300-LOC source budget, and
wires three new bloat-findings rows into the Compliance Dashboard so
the campaign's progress is operator-visible.

Why the split is thematic, not per-phase:
The original file is not actually organized per-phase. Every per-phase
behaviour lives in `tools/verifiers/*_compliance.py`; phase_quality.py
held the dispatch + finding-schema + aggregate-render plumbing. The
spec's literal `project.py / plan.py / build.py / ...` layout would
have produced 7 near-empty modules. The thematic split below maps 1:1
to the original file's natural seams.

Package layout (was 1108 LOC -> all submodules <=300 LOC each):

  shared/scripts/lib/phase_quality/
    __init__.py            154 LOC  re-exports the pre-split surface
    _constants.py          109 LOC  phase maps, category sets, paths
    _flags.py               58 LOC  SHIPWRIGHT_* env-flag readers
    _resolution.py         196 LOC  project/phase/run_id resolution
    _findings.py           272 LOC  finding schema + atomic writer
    _runners.py            255 LOC  canon/workflow/infra/trace/quality/spec
    _aggregates.py         180 LOC  loaders + locked aggregate driver
    _dashboard_render.py   213 LOC  skill-compliance MD renderers
    _bloat_findings.py     105 LOC  baseline -> dashboard summary

Public-surface preservation:
Every symbol exported from the pre-split `__all__` is re-exported by
`__init__.py`. Verified via parametrized round-trip test in
`test_collectors_dashboard.py::test_phase_quality_public_surface_re_exports`
and via the subprocess functional probe (the Stop-hook end-to-end
exits 0 with real findings emitted on the new package).

Compliance Dashboard bloat-findings column (the second half of the
sub-iterate spec):
`shared.scripts.lib.phase_quality.collect_bloat_summary(project_root)`
reads `shipwright_bloat_baseline.json`, re-measures each entry against
the worktree, and returns three counters:

  - over_limit:    measured > limit AND state != exception
  - in_allowlist:  count of state=grandfathered + state=exception
  - ratchet_delta: sum of (measured - current) across grandfathered

The compliance plugin's `compliance_report.py` consumes those via a
new sibling `_bloat_dashboard_rows.py` (extracted so the
grandfathered `compliance_report.py` stays at its 405-LOC ceiling and
does not ratchet up). Three rows render in both the events-mode and
legacy-mode Quality Indicators table:

  | Bloat over-limit | N | PASS/WARN | <diagnostic> |
  | Bloat in allowlist | N entries | INFO | |
  | Bloat ratchet delta | +/-N lines | PASS/WARN | <diagnostic> |

Real round-trip probe on the dev repo (regenerated dashboard.md):

  | Bloat over-limit    | 153         | WARN | ... |
  | Bloat in allowlist  | 153 entries | INFO |     |
  | Bloat ratchet delta | -21 lines   | PASS |     |

The non-zero over_limit count is expected at this point in Campaign B
(B3 is the third of many sub-iterates; the campaign goal is to drive
this to 0 by completion). The negative ratchet_delta confirms the
campaign is shrinking the grandfathered surface, not ratcheting up.

CLEANUP-INVARIANT compliance:
* `shipwright_bloat_baseline.json` entry for
  `shared/scripts/lib/phase_quality.py` is removed in the same commit
  (file no longer exists; it is a package directory now).
* Every NEW submodule is <=300 LOC.
* `compliance_report.py` measured = 405 (== baseline ceiling); the
  anti-ratchet rule holds.
* No new entries added to the baseline.

Test coverage:
* 19 new tests in
  `plugins/shipwright-compliance/tests/test_collectors_dashboard.py`
  covering the producer (over_limit/in_allowlist/ratchet_delta math +
  fail-open + missing-file paths) and the round-trip probe
  (`generate_file` writes the three rows to dashboard.md against a
  fixture).
* 108 pre-existing `test_audit_phase_quality* / test_phase_quality_*`
  tests still green (public surface preserved).
* 25 pre-existing `test_compliance_report` tests still green
  (legacy + events-mode rendering unchanged outside the new rows).
* `test_events_log_ssot.py` _MAIN_REPO_ONLY allowlist updated:
  `lib/phase_quality.py` -> `lib/phase_quality/_resolution.py`
  (the raw event-log join moved with the resolver helpers into the
  new submodule).
