+++ shared/tests/test_events_log_ssot.py
@@ -51,8 +51,10 @@ _MAIN_REPO_ONLY = {
         "generic verifier helper for build/adopt phase verifiers (main-repo phases)",
     "tools/verifiers/adopt_compliance.py":
         "/shipwright-adopt A7 verifier; adopt executes in the main repo",
-    "lib/phase_quality.py":
-        "Phase-Quality Stop hook; runs with cwd = main project root",
+    "lib/phase_quality/_resolution.py":
+        "Phase-Quality Stop hook resolution helpers; runs with cwd = main "
+        "project root. Iterate Campaign B (B3) split: the raw join moved "
+        "from the pre-split lib/phase_quality.py into this submodule.",
 }
 
 # A path-join onto the event log: `<expr> / EVENT_FILE` or
